+++ app/src/main/java/mx/itesm/life_tqueremos/MenuActivity.java
@@ -8,6 +8,7 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.widget.CardView;
 import android.view.MenuItem;
 import android.view.View;
 
@@ -26,6 +27,7 @@ public class MenuActivity extends AppCompatActivity implements NavigationView.On
     private DrawerLayout mDrawerLayout;
     private ActionBarDrawerToggle mToggle;
     private FirebaseAuth firebaseAuth;
+    private CardView cvEspiritual, cvEmocional;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -40,6 +42,30 @@ public class MenuActivity extends AppCompatActivity implements NavigationView.On
         mToggle.syncState();
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
+        cvEspiritual = (CardView) findViewById(R.id.cardView_espiritual);
+        cvEmocional = (CardView) findViewById(R.id.cardView_emocional);
+
+        cvEspiritual.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(MenuActivity.this, PollActivity.class);
+//                Bundle bundle = new Bundle();
+//                bundle.putString("encuesta", "Espiritual");
+                intent.putExtra("encuesta", "Espiritual");
+                startActivity(intent);
+            }
+        });
+
+        cvEmocional.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(MenuActivity.this, PollActivity.class);
+//                Bundle bundle = new Bundle();
+//                bundle.putString("encuesta", "Espiritual");
+                intent.putExtra("encuesta", "Emocional");
+                startActivity(intent);
+            }
+        });
 
         NavigationView navigationView = findViewById(R.id.navigationView);
         navigationView.setNavigationItemSelectedListener(this);
